requirements.txt
pyTelegramBotAPI==4.15.4
pillow==10.2.0
PyMySQL==1.1.0
python-dotenv==1.0.1